+++ django/django_apps/settings.py
@@ -12,26 +12,28 @@ https://docs.djangoproject.com/en/4.2/ref/settings/
 
 from pathlib import Path
 import os
-import yaml
-from logging.handlers import RotatingFileHandler
+# import json
 
-# Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
 
+# # Define the path to the local settings JSON file
+# LOCAL_SETTINGS_PATH = os.path.join(BASE_DIR, 'local_settings.json')
 
-# Quick-start development settings - unsuitable for production
-# See https://docs.djangoproject.com/en/4.2/howto/deployment/checklist/
+# # Load settings from the JSON file if it exists
+# if os.path.isfile(LOCAL_SETTINGS_PATH):
+#     with open(LOCAL_SETTINGS_PATH, 'r') as file:
+#         local_settings = json.load(file)
+#         SECRET_KEY = local_settings.get('SECRET_KEY', 'default-secret-key')
+#         DEBUG = local_settings.get('debug', False)
+#         BASE_URL = local_settings.get('base_url', 'http://localhost:8000')
+# else:
+#     raise FileNotFoundError(f'{LOCAL_SETTINGS_PATH} does not exist. Please provide a local_settings.json file.')
 
-LOCAL_SETTINGS_PATH = os.path.join(os.path.dirname(BASE_DIR), 'settings.yaml')
+DEBUG = True
 
-# Load settings from the YAML file if it exists
-if os.path.isfile(LOCAL_SETTINGS_PATH):
-    with open(LOCAL_SETTINGS_PATH, 'r') as file:
-        local_settings = yaml.safe_load(file)
-        SECRET_KEY = local_settings.get('SECRET_KEY', 'default-secret-key')
-        DEBUG = local_settings.get('DEBUG', False)
-else:
-    raise FileNotFoundError(f'{LOCAL_SETTINGS_PATH} does not exist. Please provide a settings.yaml file.')
+BASE_URL = 'https://127.0.0.1:8000'
+
+SECRET_KEY = 'safe_key'
 
 ALLOWED_HOSTS = [
   'tedisrozenfelds.pythonanywhere.com',