+++ django/changelog.md
@@ -6,3 +6,4 @@
 
 1. text-to-audio endpoint should accept text as body not quqery parameter
 2. text-to-audio endpoint should check if audio exists and if not create it
+3. Use settings from `local_settings.json` file